Parse function declaration
The textbook says at this point the following code is possible:

    fun asd() { print "zxc"; }
    print asd;

But vm.c -> run() -> case OP_RETURN immediately hits, so OP_SET_GLOBAL
for the function name is not executed. I'm not sure if the book is wrong
or I did something wrong. Let's see after implementing function call.